pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepagent-azure-cli"
version = "0.2.0"
description = "A turnkey coding assistant CLI powered by LangChain DeepAgents and Azure OpenAI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<4.0"
authors = [
    {name = "Nicolas Cravino", email = "ncravino@mac.com"},
]
keywords = ["ai", "coding-assistant", "azure-openai", "deepagents", "langgraph", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "deepagents>=0.0.20",
    "deepagents-cli>=0.0.20",
    "langchain-openai>=0.3.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "rich>=13.0.0",
    "textual>=0.40.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
search = ["tavily-python>=0.5.0"]
all = ["tavily-python>=0.5.0"]

[project.scripts]
deepagent-azure = "deepagent_azure_cli.cli:main"
daz = "deepagent_azure_cli.cli:main"

[project.urls]
Homepage = "https://github.com/ai-agents-cybersecurity/deepagent-azure-cli"
Repository = "https://github.com/ai-agents-cybersecurity/deepagent-azure-cli"
Issues = "https://github.com/ai-agents-cybersecurity/deepagent-azure-cli/issues"

[tool.setuptools.packages.find]
where = ["src"]
